Increase open file limit on the ATC
We were facing some CI trouble the past couple weeks and looked into the
logs for a root cause. We found that the ATC was complaining about
having "too many open files". There was discussion[1] which suggested to
bump this value. We chose this value as it is the Linux system
default[2].

[1]: https://discuss.concourse-ci.org/t/v4-2-1-too-many-open-files/807
[2]: https://elixir.bootlin.com/linux/latest/source/fs/file.c#L22

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>