Flow to update CIDRs for workers
Mostly to handle the Openstack IP changing on us. It was too much work
to automate the source of the CIDR, as the list is blocked from outside
the Pivotal/VMWare network.

Instead, the lastpass note is the canonical source, updated with
comments on how to obtain the CIDRs.

Removed the time interval repaving.